Refactor sshecho binary into importable tests utility
Moves `bin/e2e/sshecho` to `tests/e2e/utils/sshecho` and refactors it into a reusable package for end-to-end and unit testing, removing `main` to prevent standalone execution. Also updates the internal app initializer to export `NewApp()` function.
